Fix not being able to select an event in GoTo.
This fixes an editor issue related to assigned theme to the event selector.

Also makes sure that EditorProperty uses the new timeline format, making
use of the EventManager node instead of trying to get the timeline path
from external paths.

Adds a label under the next event property to show the timeline name.